chore(files): restructure files and folders
- should restructure files and folders appropriately

[Chore #164682064]
