+++ SERVER/test/message.test.js
@@ -1,8 +1,8 @@
 import chai from 'chai';
 import chaiHttp from 'chai-http';
-import app from '../src/controller/route';
-import users from '../src/model/Database';
-import Message from '../src/model/Message';
+import app from '../src/app';
+import users from '../src/Database/Database';
+import Message from '../src/Model/Message';
 
 const mary = users[2];
 const ayo = users[0];
@@ -11,84 +11,30 @@ chai.use(chaiHttp);
 
 const { expect } = chai;
 
-describe('route', () => {
+describe('Messages', () => {
   let obj;
 
+  describe('/auth/login', () => {
+    beforeEach(() => {
+      obj = {
+        email: 'maryj@test.com',
+        password: 'spiderman123',
+      };
+    });
+
+    it('should return status 200', (done) => {
+      chai.request(app)
+        .post('/api/v1/auth/login')
+        .send(obj)
+        .end((err, res) => {
+          authToken = res.body.data[0].token;
+          expect(res).to.have.status(200);
+          done();
+        });
+    });
+  });
+
   describe('post', () => {
-    describe('/auth/signup', () => {
-      beforeEach(() => {
-        obj = {
-          email: 'andela@andela.com',
-          firstName: 'Epic',
-          lastName: 'Andela',
-          password: 'epicTower',
-          rePassword: 'epicTower',
-        };
-      });
-
-      it('should return status 201', (done) => {
-        chai.request(app)
-          .post('/api/v1/auth/signup')
-          .send(obj)
-          .end((err, res) => {
-            expect(res).to.have.status(201);
-            done();
-          });
-      });
-
-      it('should have the following property: status, error', (done) => {
-        chai.request(app)
-          .post('/api/v1/auth/signup')
-          .send(obj)
-          .end((err, res) => {
-            expect(res.body).to.have.property('error');
-            expect(res.body).to.have.property('status');
-            done();
-          });
-      });
-
-      it('should not signup with same mail adrress more than once', (done) => {
-        chai.request(app)
-          .post('/api/v1/auth/signup')
-          .send(obj)
-          .end((err, res) => {
-            expect(res).to.have.status(400);
-            done();
-          });
-      });
-    });
-
-    describe('/auth/login', () => {
-      beforeEach(() => {
-        obj = {
-          email: 'maryj@test.com',
-          password: 'spiderman123',
-        };
-      });
-
-      it('should return status 200', (done) => {
-        chai.request(app)
-          .post('/api/v1/auth/login')
-          .send(obj)
-          .end((err, res) => {
-            authToken = res.body.data[0].token;
-            expect(res).to.have.status(200);
-            done();
-          });
-      });
-
-      it('should have the following property: data, status', (done) => {
-        chai.request(app)
-          .post('/api/v1/auth/login')
-          .send(obj)
-          .end((err, res) => {
-            expect(res.body).to.have.property('data');
-            expect(res.body).to.have.property('status');
-            done();
-          });
-      });
-    });
-
     describe('/messages', () => {
       it('should return status 201', (done) => { //  testing for saving and message...
         obj = {
@@ -183,7 +129,7 @@ describe('route', () => {
           .set('authorization', authToken)
           .end((err, res) => {
             expect(res.body.data
-              .every(message => parseInt(message.receiverId, 10) === 3))
+              .every(message => parseInt(message.receiverId, 10) === 2))
               .to.equal(true);
             done();
           });
@@ -287,6 +233,38 @@ describe('route', () => {
           .set('authorization', authToken)
           .end((err, res) => {
             expect(res.body.data[0].id).to.equal(1);
+            done();
+          });
+      });
+    });
+
+    describe('/api/v1/messages/sent ', () => {
+      it('should return status 200', (done) => {
+        chai.request(app)
+          .get('/api/v1/messages/sent')
+          .set('authorization', authToken)
+          .end((err, res) => {
+            expect(res).to.have.status(200);
+            done();
+          });
+      });
+      it('should have the following property: data, status', (done) => {
+        chai.request(app)
+          .get('/api/v1/messages/sent')
+          .set('authorization', authToken)
+          .end((err, res) => {
+            expect(res.body).to.have.property('data');
+            expect(res.body).to.have.property('status');
+            done();
+          });
+      });
+
+      it('should return only sent mail of user', (done) => {
+        chai.request(app)
+          .get('/api/v1/messages/sent')
+          .set('authorization', authToken)
+          .end((err, res) => {
+            expect(res.body.data.every(mail => mail.senderId === mary.getId())).to.equal(true);
             done();
           });
       });
