+++ UI/index.html
@@ -12,7 +12,7 @@
     <body>
         <div class="body">
             <div id="home" class="slider">
-                <img src="./img/email-background-images-1.v1%20(1).jpg">
+                <div></div>
                 <div>
                     <h1>Welcome To <span>Epic Mail</span></h1>
                     <p>Epic Mail is a mailing web app that enables user to exchange information over the internet.
